added firther security for dashboard
