+++ app/api/replies.ts
@@ -21,8 +21,8 @@ export async function getTweetsAndReplies({ username, limit }: TweetProps): Prom
         include_pinned: false,
       },
       headers: {
-        'X-RapidAPI-Key': process.env.NEXT_PUBLIC_RAPIDAPI_KEY,
-        'X-RapidAPI-Host': process.env.NEXT_PUBLIC_RAPIDAPI_HOST,
+        'X-RapidAPI-Key': process.env.RAPIDAPI_KEY,
+        'X-RapidAPI-Host': process.env.RAPIDAPI_HOST,
       },
     });
