+++ app/api/user-details.ts
@@ -29,8 +29,8 @@ export const getUserDetails = async (username:string) => {
       username: `${username}`,
     },
     headers: {
-      "X-RapidAPI-Key": process.env.NEXT_PUBLIC_RAPIDAPI_KEY,
-      "X-RapidAPI-Host": process.env.NEXT_PUBLIC_RAPIDAPI_HOST,
+      "X-RapidAPI-Key": process.env.RAPIDAPI_KEY,
+      "X-RapidAPI-Host": process.env.RAPIDAPI_HOST,
     },
   };
