+++ app/dashboard/[username]/page.tsx
@@ -1,0 +1,13 @@
+import { getUserDetails } from "@/app/api/user-details";
+
+export default async function DashboardPage({
+  params: { username },
+}: {
+  params: { username: string };
+}) {
+  const data = await getUserDetails(username);
+
+  console.log("data = ", data);
+
+  return <main>{JSON.stringify(data)}</main>;
+}
