+++ app/dashboard/form/page.tsx
@@ -1,0 +1,9 @@
+import { FeedbackForm } from "@/components/form-components/feedback.form";
+
+export default function UserExperience() {
+  return (
+    <main className="flex flex-col">
+      <FeedbackForm />
+    </main>
+  );
+}
