+++ components/cards/additional-details.card.tsx
@@ -1,6 +1,6 @@
 // Additional information part in profile section
 "use client";
-import { UserDataProps } from "@/lib/fetches/user-details";
+import { UserDataProps } from "@/app/api/user-details";
 import {
   BsFillCalendarDateFill,
   BsFillPeopleFill,
